more stable netbase (just another wrapper around, in this case a refcounter)
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1222 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/CChannel.cpp
@@ -187,7 +187,7 @@ bool CChannel::Process(CBytestream *bs)
 		Reliable.Clear();
 
 	// Check if pong has been acknowledged
-	if(SequenceAck >= iPongSequence)  {
+	if(SequenceAck >= (size_t)iPongSequence)  {
 		iPongSequence = -1;  // Ready for new pinging
 		iPing = (int)((tLX->fCurTime - fLastPingSent) * 1000);
 	}
